+++ Services/BestService.Services.Data/ICommentsService.cs
@@ -4,6 +4,6 @@
 
     public interface ICommentsService
     {
-        Task<int> CreateAsync(string content, string userId, int companyId, int rating);
+        Task<int> CreateAsync(string content, string userId, int companyId, byte rating);
     }
 }
